+++ 3rdparty/fcpp/cpp6.c
@@ -488,7 +488,7 @@ char *savestring(struct Global *global, char *text)
 }
 
 ReturnCode getfile(struct Global *global,
-                   int bufsize, /* Line or define buffer size   */
+                   size_t bufsize, /* Line or define buffer size   */
                    char *name,
                    FILEINFO **file) /* File or macro name string        */
 {
@@ -496,7 +496,7 @@ ReturnCode getfile(struct Global *global,
    * Common FILEINFO buffer initialization for a new file or macro.
    */
 
-  int size;
+  size_t size;
 
   size = strlen(name);                          /* File/macro name      */
 
@@ -583,12 +583,12 @@ DEFBUF *defendel(struct Global *global,
   char *np;
   int nhash;
   int temp;
-  int size;
+  size_t size;
 
   for (nhash = 0, np = name; *np != EOS;)
     nhash += *np++;
   size = (np - name);
-  nhash += size;
+  nhash += (int)size;
   prevp = &global->symtab[nhash % SBSIZE];
   while ((dp = *prevp) != (DEFBUF *) NULL) {
     if (dp->hash == nhash
@@ -606,7 +606,7 @@ DEFBUF *defendel(struct Global *global,
     prevp = &dp->link;
   }
   if (!delete) {
-    dp = (DEFBUF *) malloc((int) (sizeof (DEFBUF) + size));
+    dp = (DEFBUF *) malloc(sizeof (DEFBUF) + size);
     dp->link = *prevp;
     *prevp = dp;
     dp->hash = nhash;
