+++ 3rdparty/glsl-optimizer/src/glsl/glcpp/glcpp.h
@@ -50,7 +50,7 @@ typedef struct token_list token_list_t;
 
 typedef union YYSTYPE
 {
-	int64_t ival;
+	int ival;
 	char *str;
 	string_list_t *string_list;
 	token_t *token;
